pyproject.toml
[project]
authors = [{name = "ElvirKarimov", email = "karimov@airi.net"}]
name = "twai_project"
requires-python = ">= 3.11"
version = "0.1.0"
dependencies = ["datasets", "pytorch-gpu", "transformers>=4.49.0,<5", "accelerate>=1.4.0,<2", "python-dotenv>=1.0.1,<2", "transformer-lens>=2.15.0,<3", "circuitsvis>=1.43.3,<2"]

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.pixi.project]
channels = ["conda-forge"]
platforms = ["linux-64"]

[tool.pixi.pypi-dependencies]
twai_project = { path = ".", editable = true }
torch = { version = "*", index = "https://download.pytorch.org/whl/cu124" }

[tool.pixi.system-requirements]
cuda = "12.3"

[tool.pixi.tasks]

[tool.pixi.dependencies]
cuda-version = "12.3.*"
python = "==3.11"
ipython = ">=9.0.0,<10"
jupyterlab = ">=4.3.5,<5"
ipywidgets = ">=8.1.5,<9"
tqdm = ">=4.67.1,<5"
